pyproject.toml
[project]
name = "aggie"
version = "0.1.0"
description = "Privacy-first voice assistant daemon for Linux"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = {text = "GPL-3.0-or-later"}

dependencies = [
    "anthropic>=0.40.0",
    "faster-whisper>=1.0.0",
    "mcp>=1.0.0",
    "numpy>=1.24.0,<2.0.0",
    "openwakeword>=0.6.0",
    "pathvalidate>=3.3.1",
    "piper-tts>=1.2.0",
    "pyyaml>=6.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0.0",  # For detailed GPU info; ctranslate2 handles CUDA acceleration
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
aggie = "aggie.__main__:main"
aggie-ctl = "aggie.cli.ctl:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
